third_party/WebKit/LayoutTests/platform/win/virtual/display_list_2d_canvas/fast/canvas/canvas-imageSmoothingQuality-expected.txt

CONSOLE WARNING: line 44: The provided value '3223' is not a valid enum value of type ImageSmoothingQuality.
CONSOLE WARNING: line 44: The provided value 'bad_input' is not a valid enum value of type ImageSmoothingQuality.
CONSOLE WARNING: line 44: The provided value 'LOW' is not a valid enum value of type ImageSmoothingQuality.
CONSOLE WARNING: line 44: The provided value 'Medium' is not a valid enum value of type ImageSmoothingQuality.
This is a testharness.js-based test.
PASS Tests for the imageSmoothingQuality attribute. 
Harness: the test ran to completion.
